// frontend/src/shared/api/generated/employers/models/employmentType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Benefit Employers
 * OpenAPI spec version: 0.1.0
 */

export type EmploymentType = typeof EmploymentType[keyof typeof EmploymentType];


export const EmploymentType = {
  full_time: 'full_time',
  part_time: 'part_time',
  project: 'project',
  internship: 'internship',
} as const;