// frontend/src/shared/api/generated/employers/models/grade.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Benefit Employers
 * OpenAPI spec version: 0.1.0
 */

export type Grade = typeof Grade[keyof typeof Grade];


export const Grade = {
  intern: 'intern',
  junior: 'junior',
  middle: 'middle',
  senior: 'senior',
  lead: 'lead',
} as const;