// frontend/src/shared/api/generated/employers/models/iTRole.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Benefit Employers
 * OpenAPI spec version: 0.1.0
 */

export type ITRole = typeof ITRole[keyof typeof ITRole];


export const ITRole = {
  backend: 'backend',
  frontend: 'frontend',
  fullstack: 'fullstack',
  mobile: 'mobile',
  devops: 'devops',
  qa: 'qa',
  qa_automation: 'qa_automation',
  data_scientist: 'data_scientist',
  ml_engineer: 'ml_engineer',
  data_engineer: 'data_engineer',
  data_analyst: 'data_analyst',
  system_analyst: 'system_analyst',
  business_analyst: 'business_analyst',
  product_manager: 'product_manager',
  project_manager: 'project_manager',
  designer: 'designer',
  gamedev: 'gamedev',
  security: 'security',
  embedded: 'embedded',
  dba: 'dba',
  team_lead: 'team_lead',
  architect: 'architect',
} as const;