// frontend/src/shared/api/generated/employers/models/vacancyIn.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Benefit Employers
 * OpenAPI spec version: 0.1.0
 */
import type { EmploymentType } from './employmentType';
import type { Grade } from './grade';
import type { ITRole } from './iTRole';
import type { WorkFormat } from './workFormat';

export interface VacancyIn {
  salary_from?: number | null;
  salary_to?: number | null;
  /** @pattern ^[A-Z]{3}$ */
  currency?: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  title: string;
  /**
     * @minLength 1
     * @maxLength 8000
     */
  description: string;
  specialization: ITRole;
  grade: Grade;
  /**
     * @maxItems 30
     * @items.minLength 1
     * @items.maxLength 64
     */
  skills?: string[];
  work_format?: WorkFormat | null;
  employment_type?: EmploymentType | null;
  city?: string | null;
  need_id?: string | null;
}