// frontend/src/shared/api/generated/employers/models/vacancyResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Benefit Employers
 * OpenAPI spec version: 0.1.0
 */
import type { CompanyBrief } from './companyBrief';
import type { EmploymentType } from './employmentType';
import type { Grade } from './grade';
import type { ITRole } from './iTRole';
import type { VacancyStatus } from './vacancyStatus';
import type { WorkFormat } from './workFormat';

export interface VacancyResponse {
  salary_from?: number | null;
  salary_to?: number | null;
  /** @pattern ^[A-Z]{3}$ */
  currency?: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  title: string;
  /**
     * @minLength 1
     * @maxLength 8000
     */
  description: string;
  specialization: ITRole;
  grade: Grade;
  /**
     * @maxItems 30
     * @items.minLength 1
     * @items.maxLength 64
     */
  skills?: string[];
  work_format?: WorkFormat | null;
  employment_type?: EmploymentType | null;
  city?: string | null;
  need_id?: string | null;
  id: string;
  company: CompanyBrief;
  owner_id: string;
  status: VacancyStatus;
  published_at: string | null;
  created_at: string;
  updated_at: string;
}